refactor for possession strategy
