+++ src/proto/strategy.proto
@@ -1,0 +1,11 @@
+syntax = "proto3";
+
+package TbotsProto;
+
+message PossessionStrategy
+{
+    int32 goalkeepers = 1;
+    int32 defenders = 2;
+    int32 attackers = 3;
+    int32 supporters = 4;
+};
