+++ src/software/ai/hl/stp/strategy/BUILD
@@ -31,23 +31,16 @@ cc_library(
 
 cc_library(
     name = "strategy_impl",
-    srcs = ["strategy_impl.cpp"],
-    hdrs = ["strategy_impl.h"],
+    hdrs = ["strategy_impl.hpp"],
     visibility = ["//visibility:private"],
     deps = [
         ":pass_strategy",
-        ":possession_strategy",
         "//proto:tbots_cc_proto",
         "//software/ai/evaluation:calc_best_shot",
         "//software/ai/evaluation:shot",
-        "//software/ai/hl/stp/tactic/offense_support_tactics:offense_support_tactic",
+        "//software/ai/hl/stp/tactic/offense_support_tactics:offense_support_type",
         "//software/ai/passing:pass_with_rating",
         "//software/geom:pose",
         "//software/world:field",
     ],
-)
-
-cc_library(
-    name = "possession_strategy",
-    hdrs = ["possession_strategy.h"],
 )
