--- src/software/ai/hl/stp/strategy/possession_strategy.h
@@ -1,8 +1,0 @@
-#pragma once
-
-struct PossessionStrategy
-{
-    bool has_attacker;
-    int num_defenders;
-    int num_support;
-};
