+++ src/software/ai/hl/stp/strategy/strategy.cpp
@@ -1,11 +1,10 @@
 #include "software/ai/hl/stp/strategy/strategy.h"
 
-#include "software/ai/passing/eighteen_zone_pitch_division.h"
 #include "software/ai/passing/pass_evaluation.hpp"
 #include "software/ai/passing/pass_with_rating.h"
 
 Strategy::Strategy(const TbotsProto::AiConfig& ai_config, const Field& field)
-    : strategy_(std::make_shared<StrategyImpl>(ai_config, field))
+    : strategy_(std::make_shared<StrategyImpl<EighteenZoneId>>(ai_config, field))
 {
 }
 
@@ -24,7 +23,7 @@ void Strategy::updateWorld(const WorldPtr& world_ptr)
     strategy_->updateWorld(world_ptr);
 }
 
-std::shared_ptr<StrategyImpl> Strategy::operator->()
+std::shared_ptr<StrategyImpl<EighteenZoneId>> Strategy::operator->()
 {
     CHECK(strategy_->hasWorld())
         << "[StrategyImpl] Cannot generate next Strategy without a World!";
