+++ src/software/ai/hl/stp/tactic/offense_support_tactics/offense_support_tactic.h
@@ -1,9 +1,8 @@
 #pragma once
 
+#include "software/ai/hl/stp/tactic/offense/offense_support_type.h"
 #include "software/ai/hl/stp/tactic/tactic.h"
-#include "software/util/make_enum/make_enum.h"
-
-MAKE_ENUM(OffenseSupportType, PASS_RECEIVER, SUPPORT);
+#include "software/ai/hl/stp/strategy/strategy.h"
 
 class OffenseSupportTactic : public Tactic
 {
