+++ src/software/ai/hl/stp/tactic/offense_support_tactics/offense_support_type.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "software/util/make_enum/make_enum.h"
+
+MAKE_ENUM(OffenseSupportType, PASS_RECEIVER, SUPPORT);
